changed 'unsigned int' variables' type to size_t.
git-svn-id: http://alma-dms/svn/sakura/trunk@481 9a79bd28-2a0b-46f8-beb7-14fc758f186d

+++ libsakura/src/baseline.cc
@@ -20,11 +20,11 @@ using ::Eigen::Aligned;
 namespace {
 
 inline void GetBaselineModel(
-		size_t num_chan, unsigned int order, double *out) {
+		size_t num_chan, size_t order, double *out) {
 	assert(LIBSAKURA_SYMBOL(IsAligned)(out));
-	unsigned int num_model = order + 1;
+	size_t num_model = order + 1;
 
-	for (unsigned int i = 0; i < num_model; ++i) {
+	for (size_t i = 0; i < num_model; ++i) {
 		for (size_t j = 0; j < num_chan; ++j) {
 			out[num_chan*i+j] = pow(static_cast<double>(j), static_cast<double>(i));
 		}
@@ -32,8 +32,8 @@ inline void GetBaselineModel(
 }
 
 inline void DoSubtractBaseline(
-		size_t num_chan, float const *in_data, bool const *in_mask, unsigned int num_model,
-		double const *model_data, float clipping_threshold_sigma, unsigned int num_fitting_max,
+		size_t num_chan, float const *in_data, bool const *in_mask, size_t num_model,
+		double const *model_data, float clipping_threshold_sigma, size_t num_fitting_max,
 		bool get_residual, float *out) {
 	//std::cout << "DoSubtractBaselineEigen function is called" << std::endl;
 
@@ -52,7 +52,7 @@ inline void DoSubtractBaseline(
 	float *residual_data = reinterpret_cast<float *>(LIBSAKURA_PREFIX::Memory::Allocate(sizeof(float)*num_chan));
 	//bool *new_clip_mask = reinterpret_cast<bool *>(LIBSAKURA_PREFIX::Memory::Allocate(sizeof(bool)*num_chan));
 
-	for (unsigned int i = 0; i < num_fitting_max; ++i) {
+	for (size_t i = 0; i < num_fitting_max; ++i) {
 		LIBSAKURA_SYMBOL(OperateLogicalAnd)(num_chan, in_mask, clip_mask, composite_mask);
 		LIBSAKURA_SYMBOL(GetBestFitModel)(num_chan, in_data, composite_mask, num_model, model_data, best_fit_model);
 		LIBSAKURA_SYMBOL(OperateFloatSubtraction)(num_chan, in_data, best_fit_model, residual_data);
@@ -98,10 +98,10 @@ inline void DoSubtractBaseline(
 
 inline void SubtractBaselinePolynomial(
 		size_t num_chan, float const *in_data, bool const *in_mask,
-		unsigned int order, float clipping_threshold_sigma,
-		unsigned int num_fitting_max, bool get_residual,
+		size_t order, float clipping_threshold_sigma,
+		size_t num_fitting_max, bool get_residual,
 		float *out) {
-	unsigned int num_model = order + 1;
+	size_t num_model = order + 1;
 	double *model_data = reinterpret_cast<double *>(LIBSAKURA_PREFIX::Memory::Allocate(sizeof(double)*num_chan*num_model));
 
 	GetBaselineModel(num_chan, order, model_data);
@@ -116,23 +116,23 @@ inline void SubtractBaselinePolynomial(
 namespace LIBSAKURA_PREFIX {
 void ADDSUFFIX(Baseline, ARCH_SUFFIX)::SubtractBaselinePolynomial(
 		size_t num_chan, float const in_data[/*num_chan*/],
-		bool const in_mask[/*num_chan*/], unsigned int order,
-		float clipping_threshold_sigma, unsigned int num_fitting_max,
+		bool const in_mask[/*num_chan*/], size_t order,
+		float clipping_threshold_sigma, size_t num_fitting_max,
 		bool get_residual, float out[/*num_chan*/]) const {
 	SubtractBaselinePolynomial(num_chan, in_data, in_mask, order,
 			clipping_threshold_sigma, num_fitting_max, get_residual, out);
 }
 
 void ADDSUFFIX(Baseline, ARCH_SUFFIX)::GetBaselineModel(
-		size_t num_chan, unsigned int order, double out[/*(order+1)*num_chan*/]) const {
+		size_t num_chan, size_t order, double out[/*(order+1)*num_chan*/]) const {
 	GetBaselineModel(num_chan, order, out);
 }
 
 void ADDSUFFIX(Baseline, ARCH_SUFFIX)::DoSubtractBaseline(
 		size_t num_chan, float const in_data[/*num_chan*/],
-		bool const in_mask[/*num_chan*/], unsigned int num_model,
+		bool const in_mask[/*num_chan*/], size_t num_model,
 		double const model_data[/*num_model * num_chan*/],
-		float clipping_threshold_sigma, unsigned int num_fitting_max,
+		float clipping_threshold_sigma, size_t num_fitting_max,
 		bool get_residual, float out[/*num_chan*/]) const {
 	DoSubtractBaseline(num_chan, in_data, in_mask, num_model, model_data,
 		clipping_threshold_sigma, num_fitting_max, get_residual, out);
